solana_exec

Function make_swap_instructions

pub fn make_swap_instructions(
    swap_configs: &Vec<SwapConfig>,
    central_context: &Arc<CentralContext>,
    compute_unit_budget: &mut u32,
) -> Vec<Instruction>

Create swap instructions from a list of swap configs. Processes swap configs in order and creates protocol-specific swap instructions. Also handles opening token accounts before swaps and closing them after. Updates the compute unit budget based on operations performed.